Add a validation for the movie's imdbId
There was a missing validation to get a movie's imdbId. It was only checking if there was an anchor with a href attribute, but not if the href attribute contained 'imdb'. So if there was no imdb anchor, the regex was still called, but on an undefined value.

Update the tests movies to work, since the movies in the tests are no longer in theaters (may need to find a permanent solution).

Added 2 tests to test the 2 validations : theathers not found & movie not found.